Refactoring and adjustments based on David's comments.
