+++ srcalc/src/main/java/gov/va/med/srcalc/vista/RpcVistaPatientDao.java
@@ -157,18 +157,18 @@ public class RpcVistaPatientDao implements VistaPatientDao
     
     private void parseWeightResults(final Patient patient, final List<String> weightResults) throws ParseException
     {
-    	/* The last entries are the most recent so we use those.
-    	* Get the most recent weight measurement within the already specified range.
-    	* The format expected from VistA is:
-    	*   21557^04/17/09@12:00   Wt:   185.00 lb (84.09 kg)  _NURSE,ONE
-        *          @12:00   Body Mass Index:   25.86
-        *   22296^08/24/09@14:00   Wt:   190.00 lb (86.36 kg)  _NURSE,ONE
-        *          @14:00   Body Mass Index:   26.56
-    	* Where the most recent weight is the last result and each result consists of two lines.
-    	* The first line is a measurement identifier, the date and time, the weight in pounds and kilograms
-    	* and the person providing the measurement. The second line is the time on the same date as the
-    	* weight measurement, along with the BMI for the patient.
-        */ 
+        /*-
+         * The last entries are the most recent so we use those. Get the most recent weight measurement within the
+         * already specified range. The format expected from VistA is:
+         * 21557^04/17/09@12:00 Wt: 185.00 lb (84.09 kg) _NURSE,ONE
+         *        @12:00 Body Mass Index: 25.86
+         * 22296^08/24/09@14:00 Wt: 190.00 lb (86.36 kg) _NURSE,ONE
+         *        @14:00 Body Mass Index: 26.56
+         * Where the most recent weight is the last result and each result consists of two
+         * lines. The first line is a measurement identifier, the date and time, the weight in pounds and kilograms and
+         * the person providing the measurement. The second line is the time on the same date as the weight measurement,
+         * along with the BMI for the patient.
+         */
         final List<String> weightLineTokens = Splitter.on(Pattern.compile("[\\s\\^]+"))
                 .splitToList(weightResults.get(weightResults.size()-2));
     	// Get the date of the measurement
@@ -230,7 +230,7 @@ public class RpcVistaPatientDao implements VistaPatientDao
             {
                 // If an exception occurs for any reason, move to the next lab so that as much patient
                 // data as possible can still be retrieved.
-                fLogger.debug("Unable to retrieve lab {}. {}", labRetrievalEnum.name(), e.toString());
+                fLogger.warn("Unable to retrieve lab {}. {}", labRetrievalEnum.name(), e.toString());
             }
         }
     }
