+++ srcalc/src/test/java/gov/va/med/srcalc/web/controller/CalculationControllerIT.java
@@ -136,7 +136,7 @@ public class CalculationControllerIT extends IntegrationTest
         varParams.add("preopPneumonia", "true");
         varParams.add("alkalinePhosphatase", ">125mU/ml");
         varParams.add("bun", VariableEntry.SPECIAL_NUMERICAL);
-        varParams.add("bun" + VariableEntry.SEPARATOR + VariableEntry.SPECIAL_NUMERICAL, "15.0");
+        varParams.add(VariableEntry.makeNumericalInputName("bun"), "15.0");
         
         final MockHttpServletRequestBuilder request =
                 post("/enterVars").session(fSession);
@@ -220,7 +220,7 @@ public class CalculationControllerIT extends IntegrationTest
         varParams.add("preopPneumonia", "true");
         varParams.add("alkalinePhosphatase", ">125mU/ml");
         varParams.add("bun", VariableEntry.SPECIAL_NUMERICAL);
-        varParams.add("bun" + VariableEntry.SEPARATOR + VariableEntry.SPECIAL_NUMERICAL, "15.0");
+        varParams.add(VariableEntry.makeNumericalInputName("bun"), "15.0");
         
         final MockHttpServletRequestBuilder request =
                 post("/enterVars").session(fSession);
